requirements.txt
torch
torchvision
torchaudio
torchsummary
numpy
matplotlib
tqdm
pillow
opencv-python
albumentations
torchmetrics
onnx
onnxruntime
opencv-python-headless
scikit-learn
pandas
wandb
pdf2image